human_detection_feature/requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0
numpy>=1.21.0
tensorflow>=2.10.0
keras>=2.10.0
Pillow>=9.0.0
deepface>=0.0.79
mtcnn>=0.1.1
pyaudio>=0.2.11
wave>=0.0.2
SpeechRecognition>=3.10.0
pyttsx3>=2.90 